requirements.txt
altgraph==0.17.3
future==0.18.2
pefile==2022.5.30
Pillow==9.4.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
PyPDF2==3.0.1
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pywin32-ctypes==0.2.0
